+++ packages/aragon-cli/src/commands/dao_cmds/utils/getRepoTask.js
@@ -6,7 +6,7 @@ module.exports = {
   args: yargs => {
     return yargs
       .option('apmRepo', {
-        describe: 'Name of the APM repo',
+        describe: 'Name of the aragonPM repo',
       })
       .option('apmRepoVersion', {
         describe: 'Version of the package upgrading to',
@@ -35,7 +35,7 @@ module.exports = {
         // TODO: load ipfs aragon node and fetch repo again. If this time we have the artifact then return otherwise throw Error.
 
         throw new Error(
-          'Cannot find artifacts in APM repo. Please make sure the package is published and IPFS or your HTTP server running.'
+          'Cannot find artifacts in aragonPM repo. Please make sure the package is published and IPFS or your HTTP server running.'
         )
       }
     }
